test_sandbox/requirements.txt
fastapi>=0.124.2
uvicorn>=0.38.0
aiogram>=3.24.0
redis>=7.1.0
requests>=2.32.5
python-dotenv>=1.2.1
python-multipart>=0.0.21